db_util moved to database module
