+++ OpenOrchestrator/orchestrator/popups/credential_popup.py
@@ -8,7 +8,7 @@
 import tkinter
 from tkinter import ttk, messagebox
 
-from OpenOrchestrator.common import db_util
+from OpenOrchestrator.database import db_util
 
 def show_popup(name=None, username=None):
     """Creates and shows a popup to create a new credential.
