+++ OpenOrchestrator/orchestrator/popups/email_trigger_popup.py
@@ -7,7 +7,7 @@
 import tkinter
 from tkinter import ttk, messagebox
 
-from OpenOrchestrator.common import db_util
+from OpenOrchestrator.database import db_util
 
 def show_popup():
     """Creates and shows a popup to create a new email trigger.
