+++ OpenOrchestrator/scheduler/run_tab.py
@@ -5,7 +5,8 @@ import tkinter
 from tkinter import ttk
 import sys
 
-from OpenOrchestrator.common import db_util, crypto_util
+from OpenOrchestrator.common import crypto_util
+from OpenOrchestrator.database import db_util
 from OpenOrchestrator.scheduler import runner
 
 def create_tab(parent: ttk.Notebook, app) -> ttk.Frame:
